+++ augur/io/json.py
@@ -32,10 +32,16 @@ The LICENSE file included in ID3C's repo is copied below verbatim::
     SOFTWARE.
 """
 import json
+import numpy as np
+import os
+import pandas as pd
+from collections import OrderedDict
 from datetime import date, datetime, time, timedelta
+from io import RawIOBase
 from isodate import duration_isoformat
 from typing import Iterable
 from uuid import UUID
+from augur.io.file import open_file
 
 
 def as_json(value):
@@ -80,6 +86,103 @@ def load_json(value):
         return json.loads(value)
     except json.JSONDecodeError as e:
         raise JSONDecodeError(e) from e
+
+
+MINIFY_THRESHOLD_MB = 5
+
+
+def write_json(data, file, minify=None, minify_threshold_mb=MINIFY_THRESHOLD_MB, indent=2):
+    """
+    Write ``data`` as JSON to the given ``file``, creating parent directories
+    if necessary.
+
+    Parameters
+    ----------
+    data : dict
+        data to write out to JSON
+    file
+        file path or handle to write to
+    minify : bool or None, optional
+        Control output minification. ``True`` forces minified output, ``False``
+        forces non-minified output, ``None`` (default) uses auto-detection based
+        on *minify_threshold_mb*.
+        A truthy value in the environment variable :envvar:`AUGUR_MINIFY_JSON`
+        also forces minified output.
+    minify_threshold_mb : int or float, optional
+        Threshold in megabytes above which output is automatically minified.
+        Only applies when *minify* is None.
+    indent : int or None, optional
+        JSON indentation level when not minifying.
+
+    Raises
+    ------
+    OSError
+    """
+    if isinstance(file, (str, os.PathLike)):
+        #in case parent folder does not exist yet
+        parent_directory = os.path.dirname(file)
+        if parent_directory and not os.path.exists(parent_directory):
+            try:
+                os.makedirs(parent_directory)
+            except OSError: #Guard against race condition
+                if not os.path.isdir(parent_directory):
+                    raise
+
+    # Should output be minified?
+    # Order of precedence:
+    # 1. 'minify' parameter
+    # 2. 'AUGUR_MINIFY_JSON' environment variable
+    # 3. Automatically determine based on size of data
+    if minify is True:
+        effective_indent = None
+    elif minify is False:
+        effective_indent = indent
+    elif os.environ.get("AUGUR_MINIFY_JSON"):
+        effective_indent = None
+    elif json_size(data) > minify_threshold_mb * 10**6:
+        effective_indent = None
+    else:
+        effective_indent = indent
+
+    with open_file(file, 'w', encoding='utf-8') as handle:
+        sort_keys = False if isinstance(data, OrderedDict) else True
+        json.dump(data, handle, indent=effective_indent, sort_keys=sort_keys, cls=AugurJSONEncoder)
+
+
+class AugurJSONEncoder(json.JSONEncoder):
+    """
+    A custom JSONEncoder subclass to serialize data types used for various data
+    stored in dictionary format.
+    """
+    def default(self, obj):
+        if isinstance(obj, np.integer):
+            return int(obj)
+        if isinstance(obj, np.floating):
+            return float(obj)
+        if isinstance(obj, np.ndarray):
+            return obj.tolist()
+        if isinstance(obj, pd.Series):
+            return obj.tolist()
+        return super().default(obj)
+
+
+class BytesWrittenCounterIO(RawIOBase):
+    """Binary stream to count the number of bytes sent via write()."""
+    def __init__(self):
+        self.written = 0
+        """Number of bytes written."""
+
+    def write(self, b):
+        n = len(b)
+        self.written += n
+        return n
+
+
+def json_size(data):
+    """Return size in bytes of a Python object in JSON string form."""
+    with BytesWrittenCounterIO() as counter:
+        write_json(data, counter, minify=False)
+    return counter.written
 
 
 def dump_ndjson(iterable: Iterable) -> None: